'''HACK''' Steve's Super Ugly Hacked Ping Detector. I was very very bored, so I wrote this. It is EXTREMELY ugly, so don't even look at it! It uses channel 0 to trigger since this is the best working sensor. It splits the dataset into blocks. Each block is approx 2x the length of expected ping, so 2000 samples. It finds the maximum of each block. It then divides each block's maximum by the previous one, and finds where this ratio is the biggest. The ping must begin in this block.
It then applies the exact same algorithm to just that block, only the 
new 'mini-blocks' are much smaller. They are 50 samples because each 
cycle is approx 20 cycles and 50 is approx 2 cycles. Again, it sees 
where the maximum of the 'mini-blocks' increased the most. This must be 
the beginning of the ping. 

It goes a few mini-blocks in, to allow the sensors to start vibrating, 
and then plots a few mini-blocks of data from that point.

I only tested it on the five rotated datasets I have. It'll probably 
break on other data.