'''HACK HACK''' More boredom and writing ugly ugly matlab code. Okay, so finding the maximum SNR of the DFT worked surprisingly poorly, so I tried this instead. It finds the maximum of the DFT for the frequency we want. The problem was that the maximum of the DFT does not usually occur at the beginning of the ping, which is where the data is cleanest. So, I say, finding the maximum of the DFT puts us someplace within the ping that we want. Ie, if there are two pings from two frequencies, this technique will identify the ping we are looking for.
Once we are in the ping we need, we need to find where this ping actually begins. So, backtrack one full ping length 
from the max DFT trigger point, and begin searching forward using the miniblock method in the time domain. 

This is what this file does. It is ugly ugly, but it works on my seven (big whoop) very similar data sets.

Oh and we start searching for max DFT only one full ping length into the file. That way if the file starts in 
mid-ping, we don't walk off the left edge, but rather look for the second ping of this frequency.

I did this with the sliding DFT thingy in mind, but this thing actually slides 8 samples forward at a time. This is 
because matlab doesn't really have a sliding dft, so you have to get a non-sliding one and then slide it yourself. 
This is really really slow and wasteful, and going one sample at a time makes my computer very slow... so I did 8 at 
a time and I figure it is a close enough approximation of what will actually happen. The backtracker thingy runs in 
block increments anyway, so even if the trigger point isn't exactly the DFT max, it doesn't hurt it.


Sleep is now