'''Distro Board''' Fixed the DVL_INSIDE flag to do exactly what it should, several small comment and syntax changes.
The DVL_INSIDE flag now sets whether or not the bus command DVL_ON and DVL_OFF work, as well as
the drop marker commands.  It also properly sets up the input/output of the pins based on that
flag.  Also fixed it so that we no longer throw any warnings when we compile!